+++ lib/imessage.rb
@@ -1,5 +1,6 @@
-require 'imessage/version'
-require 'imessage/sender'
+require_relative 'imessage/version'
+require_relative 'imessage/parser'
+require_relative 'imessage/sender'
 
 module Imessage
 end